// src/internal/model/createUserModel.ts
/**
 * Juno
 * Juno Public API Docs
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateUserModel {
  /**
   * User email
   */
  'email': string;
  /**
   * User name
   */
  'name': string;
  /**
   * User password
   */
  'password': string;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: 'email',
      baseName: 'email',
      type: 'string',
    },
    {
      name: 'name',
      baseName: 'name',
      type: 'string',
    },
    {
      name: 'password',
      baseName: 'password',
      type: 'string',
    },
  ];

  static getAttributeTypeMap() {
    return CreateUserModel.attributeTypeMap;
  }
}